Refactor gel configure command
